refactor: modularize recurringEventHelpers utilities
