+++ docs/docs/auto-docs/utilities/recurringEvent/calculation/functions/applyRecurrenceOverrides.md
@@ -6,7 +6,7 @@
 
 > **applyRecurrenceOverrides**(`newStartAt`, `originalRecurrence`, `inputRecurrence?`): `object`
 
-Defined in: [src/utilities/recurringEventHelpers.ts:340](https://github.com/PalisadoesFoundation/talawa-api/tree/mainsrc/utilities/recurringEventHelpers.ts#L340)
+Defined in: [src/utilities/recurringEvent/calculation.ts:180](https://github.com/PalisadoesFoundation/talawa-api/tree/mainsrc/utilities/recurringEvent/calculation.ts#L180)
 
 Applies calendar-style override logic for recurring event updates.
 This function implements the following rules:
